rename: hooks 폴더 구조 변경
- hooks > queries 폴더 생성
